MAVEN-63 Invoice total tax -> BigDecimal and validation minPrice fixed

+++ src/main/java/com/cydeo/accounting_app/dto/InvoiceDTO.java
@@ -28,7 +28,7 @@ public class InvoiceDTO {
     @NotNull()
     private ClientVendorDTO clientVendor;
     private BigDecimal price;
-    private Integer tax;
+    private BigDecimal tax;
     private BigDecimal total;
 
 }
